Adds a Jenkinsfile to build automatically
We've forked audited as we're trying to run/test FreeAgent against Edge
rails. Given that the initial project is a hackday project, we haven't
got time to investigate if we can fix upstream.

We need to add a Jenkinsfile so that we can rely on this in our
FreeAgent gemfile and allow it to be built successfully.

We're install postgresql as a hack, since the audited gem requires it
for testing.